Refactor filtering and ordering functionality. Create base builders and use them for LINQ and Elasticsearch query builders.
